requirements.txt
aiohttp
requests
python-telegram-bot
telethon
cryptg
pyaes
pilk
ffmpeg-python
Pillow